Add more functions
In particular there is now a percentage function that calculates the
number / 100 of a number in case no operation has been executed previously.
In case the percentage is part of an operation the function returns the
result of the operation after calculating the percentage (i.e. 80 - 50% = 40).

Another feature added is positive/negative button, for defining if a number
is positive or negative.
The calculator can perform calculations with positive or negative numbers,
however it doesn't still work with concatenation of calculations.

Recap:
- percentage        Done
- dot               Done
- cancel all        Done
- operators anomaly Done (same anomaly of previous commit)
- positive/negative Done (but there are some bugs)